requirements.txt
pygame-ce>=2.4.0
pymunk>=6.6.0
